Learn route groups, streaming pages, and component grouping with Suspense
